Move `TestOpenAIBatchCompletionService` to a dedicated file
The `TestOpenAIBatchCompletionService` test class was moved from `test_llm_service.py` to a new file, `test_llm_batch_service.py`. This improves modularity by isolating tests for batch-related services into their own file, making the test suite easier to manage and maintain.
